# jaxrs-spec-generated/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "SOTE API",
    "description" : "SOTE API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
...
    "/events" : {
      "get" : {
        "tags" : [ "Events" ],
        "summary" : "Lists events",
        "description" : "Lists events\n",
        "operationId" : "listEvents",
        "parameters" : [ {
          "name" : "endsAfter",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "startsBefore",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "categorySlug",
          "in" : "query",
          "description" : "Filter by category slug",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a list of events",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{eventId}" : {
      "get" : {
        "tags" : [ "Events" ],
        "summary" : "Finds event",
        "description" : "Finds single event\n",
        "operationId" : "findEvent",
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "description" : "event id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a single event",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Event"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/emergencyCongestionStatuses" : {
      "get" : {
        "tags" : [ "EmergencyCongestionStatuses" ],
        "summary" : "Lists emergency congestion statuses",
        "description" : "List of emergency congestion statuses",
        "operationId" : "listEmergencyCongestionStatuses",
        "parameters" : [ {
          "name" : "sort",
          "in" : "query",
          "description" : "Sort results. See EmergencyCongestionStatusListSort for sort options",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a lists emergency congestion statuses",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EmergencyCongestionStatus"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/reviewProducts" : {
      "get" : {
        "tags" : [ "ReviewProducts" ],
        "summary" : "Lists reviewProducts",
        "description" : "Lists reviewProducts\n",
        "operationId" : "listReviewProducts",
        "responses" : {
          "200" : {
            "description" : "Returns a list of reviewProducts",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReviewProduct"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/reviewProducts/{reviewProductId}" : {
      "get" : {
        "tags" : [ "ReviewProducts" ],
        "summary" : "Finds reviewProduct",
        "description" : "Finds single reviewProduct\n",
        "operationId" : "findReviewProduct",
        "parameters" : [ {
          "name" : "reviewProductId",
          "in" : "path",
          "description" : "reviewProduct id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a single reviewProduct",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewProduct"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/reviews/{reviewId}" : {
      "get" : {
        "tags" : [ "Reviews" ],
        "summary" : "Finds review",
        "description" : "Finds single review\n",
        "operationId" : "findReview",
        "parameters" : [ {
          "name" : "reviewId",
          "in" : "path",
          "description" : "review id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a single review",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Review"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "Event" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string",
            "description" : "Category slug"
          },
          "title" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LocalizedValue"
            }
          },
          "description" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LocalizedValue"
            }
          },
          "startDate" : {
            "type" : "string",
            "description" : "Event's start date.",
            "format" : "date"
          },
          "startTime" : {
            "type" : "string",
            "description" : "Event's start time.",
            "format" : "time"
          },
          "endDate" : {
            "type" : "string",
            "description" : "Event's end date.",
            "format" : "date"
          },
          "endTime" : {
            "type" : "string",
            "description" : "Event's end time.",
            "format" : "time"
          },
          "allDay" : {
            "type" : "boolean"
          }
        }
      },
...
      "Review" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "productId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "rating" : {
            "type" : "integer",
            "format" : "int32"
          },
          "created" : {
            "type" : "string",
            "description" : "Create time.",
            "format" : "date-time"
          },
          "modified" : {
            "type" : "string",
            "description" : "Create time.",
            "format" : "date-time"
          },
          "review" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          }
        }
      },
      "EmergencyCongestionStatus" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "value" : {
            "type" : "integer",
            "format" : "int32"
          },
          "created" : {
            "type" : "string",
            "description" : "Create time.",
            "format" : "date-time"
          }
        }
      },
      "LocalizedValue" : {
        "type" : "object",
        "properties" : {
          "language" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
      "ReviewProduct" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "InternalServerError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Forbidden" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "BadRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}